divide files into smaller pieces
