+++ .size-limit.js
@@ -2,12 +2,12 @@ module.exports = [
   {
     path: ['dist/es2015/entrypoints/index.js', 'dist/es2015/entrypoints/boot.js'],
     ignore: ['tslib'],
-    limit: '3.7 KB',
+    limit: '3.9 KB',
   },
   {
     path: 'dist/es2015/entrypoints/index.js',
     ignore: ['tslib'],
-    limit: '3.5 KB',
+    limit: '3.6 KB',
   },
   {
     path: 'dist/es2015/entrypoints/boot.js',
