+++ src/entrypoints/server.ts
@@ -1,6 +1,7 @@
 import { setConfiguration } from '../configuration/config';
 import { getMarkedChunks, getMarkedFileNames } from '../loadable/markerMapper';
-import { createLoadableStream, drainHydrateMarks, printDrainHydrateMarks } from '../loadable/marks';
+import { drainHydrateMarks, printDrainHydrateMarks } from '../loadable/marks';
+import { createLoadableStream } from '../loadable/stream';
 import { getLoadableTrackerCallback } from '../trackers/globalTracker';
 import { createLoadableTransformer } from '../transformers/loadableTransformer';
 import { ImportedStream } from '../ui/context';
