+++ src/ui/context.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { defaultStream } from '../loadable/marks';
+import { defaultStream } from '../loadable/stream';
 import { Stream } from '../types';
 
 interface TakeProps {
